requirements.txt
atlassian-python-api==3.32.*
revChatGPT==1.0.*